Add files via upload
This is an encryption experiment. I can't find if there is already a name for this type of cipher (though it is close to being a polyalphabetic cipher, from what I've read), but I call it a positional substitution cipher. This cipher is made from a rule of how the rule of a substitution cipher should change depending on a character's position in a message, and the rule of the cipher also changes depending on the length of a message. My idea is that this should make it impossible to break the cipher just by counting the frequency of characters. Also adding even a single character anywhere in a message should completely change the entire encryption of it, since the length of a message determines how many times the positional substitution algorithm is performed. 
There is plenty of room for improvement, however. The algorithm is that the first letter of a message is substituted to the character 1 above it, then the second character is substituted for the character 2 above that, then the third character is substituted for the character that is 3 above. This is then done N times (where N is the length of the message). A flaw is that this is equivalent to just substituting the first character for the character that is N above it, and the second character for the character 2N above that, and so on. There are 66 characters in my alphabet, it is the standard english alphabet plus numerals, punctuation, and other commonly used symbols. So another flaw in this cipher is that no matter how long a message is, there are a maximum of 66 substitution rules that repeat in order.
This is just for fun, but any comments or suggestions would be appreciated.